Centralize common selenium configuration
Allow for use of Sauce Labs

Use Sauce Labs when env var: `SAUCE_USERNAME`
Start browser headless when env var `TRAVIS` or` HEADLESS` is set
Fall back to graphical chrome browser (default)

This functions the same as the previous version unless `SAUCE_USERNAME` is set
